edits to function of prog4 and prog4ai

+++ CSC 1300 Programs/program4ai/prog4.h
@@ -17,8 +17,8 @@ struct Cost {
 struct Monster {
     string name;
     string description;
-    int weightInPounds;
-    int heightInFeet;
+    double weightInPounds;
+    double heightInFeet;
     string location;
     int dangerLevel;
     Cost cost;
